refactoring-comment-component styles
